edit proyek

+++ src/main/resources/schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE lokasi (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id INT AUTO_INCREMENT PRIMARY KEY,
     nama_lokasi VARCHAR(255),
     negara VARCHAR(255),
     provinsi VARCHAR(255),
@@ -7,7 +7,7 @@ CREATE TABLE lokasi (
 );
 
 CREATE TABLE proyek (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id INT AUTO_INCREMENT PRIMARY KEY,
     nama_proyek VARCHAR(255),
     client VARCHAR(255),
     tgl_mulai TIMESTAMP,
@@ -17,8 +17,8 @@ CREATE TABLE proyek (
 );
 
 CREATE TABLE proyek_lokasi (
-    proyek_id BIGINT,
-    lokasi_id BIGINT,
+    proyek_id INT,
+    lokasi_id INT,
     PRIMARY KEY (proyek_id, lokasi_id),
     FOREIGN KEY (proyek_id) REFERENCES proyek(id),
     FOREIGN KEY (lokasi_id) REFERENCES lokasi(id)
